+++ src/main/java/pl/clockworkjava/domain/reservation/ReservationService.java
@@ -29,7 +29,7 @@ public class ReservationService {
         return instance;
     }
 
-    public Reservation createNewReservation(LocalDate from, LocalDate to, int roomId, int guestId) throws IllegalArgumentException {
+    public Reservation createNewReservation(LocalDate from, LocalDate to, long roomId, long guestId) throws IllegalArgumentException {
 
         //TODO: handle null room
         Room room = this.roomService.getRoomById(roomId);
@@ -69,7 +69,7 @@ public class ReservationService {
         return result;
     }
 
-    public void removeReservation(int id) {
+    public void removeReservation(long id) {
         this.repository.remove(id);
     }
 }
