+++ src/main/java/pl/clockworkjava/domain/room/RoomService.java
@@ -109,7 +109,7 @@ public class RoomService {
         this.repository.remove(id);
     }
 
-    public Room getRoomById(int roomId) {
+    public Room getRoomById(long roomId) {
         return this.repository.getById(roomId);
     }
