+++ src/main/java/pl/clockworkjava/ui/gui/reservations/AddNewReservationScene.java
@@ -79,7 +79,7 @@ public class AddNewReservationScene {
             LocalDate from = fromDateField.getValue();
             LocalDate to = toDateField.getValue();
             int guestId = guestField.getValue().getId();
-            int roomId = roomField.getValue().getId();
+            long roomId = roomField.getValue().getId();
 
             try {
                 this.reservationService.createNewReservation(from, to, roomId, guestId);
